node/requirements.txt
adafruit-circuitpython-dht
pika
python-dotenv